refactor: externalize AdvancedModeToggleButton component to own file
